REFACTORY events - Create events module at PyGrid Node App - events.py -> events/__init__.py - event_routes -> syft_events.py, model_events.py, control_events.py
